requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coverage==7.2.5
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
oauthlib==3.2.2
outcome==1.2.0
Pillow==10.0.0
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
selenium==4.10.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==1.26.15
wsproto==1.2.0
